修改gaia的头部中，长度字段为2字节

+++ EarBuds/apps/libs/gaia/gaia_voice_assistant.c
@@ -249,7 +249,7 @@ DESCRIPTION
     Handle a Voice Assistant protocol command or return FALSE if we can't
 */
 bool gaiaHandleVoiceAssistantCommand(gaia_transport *transport, uint16 command_id,
-                                      uint8 payload_length, uint8 *payload)
+                                     uint16 payload_length, uint8 *payload)
 {
     switch (command_id)
     {
